refactor(app): Refactor imports and paths
