+++ src/components/Icons/ssr/StatusIcon.tsx
@@ -1,8 +1,8 @@
+import { LoadingBIcon } from "@components";
+import { STATUSES } from "@constants";
 import { Check, Warning } from "@phosphor-icons/react/dist/ssr";
 import type { Status } from "@utils";
 import { twMerge } from "tailwind-merge";
-import { LoadingBIcon } from "../..";
-import { STATUSES } from "../../../consts";
 
 const StatusIcon = ({
 	status,
