--- src/components/Input/Slider.module.sass
@@ -1,10 +1,0 @@
-input[type="range"]
-  &::-webkit-slider-thumb
-    appearance: none
-    width: 2px
-    height: 10px
-    border-radius: 2px
-    @apply bg-white-100
-  &:disabled
-    &::-webkit-slider-thumb
-      @apply bg-transparent
