+++ src/constants/variants.ts
