+++ src/stories/Badge/Badge.stories.ts
@@ -1,7 +1,7 @@
 import { Badge } from "@components";
+import { testStatus } from "@mocks";
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, within } from "@storybook/test";
-import { testStatus } from "../mocks";
 
 const meta = {
 	title: "Base Components/Badge/Badge",
