+++ src/stories/Button.stories.tsx
@@ -1,8 +1,8 @@
 import { ArrowLineDownIcon, Button, FourLeafCloverIcon } from "@components";
-import { SIZES } from "@consts";
+import { SIZES } from "@constants";
+import { buttonVariantControl, iconControl, sizeControl } from "@mocks";
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, fn, within } from "@storybook/test";
-import { buttonVariantControl, iconControl, sizeControl } from "./mocks";
 
 const clickHandler = fn();
