+++ src/stories/Inputs/TextArea.stories.tsx
@@ -1,9 +1,9 @@
 import { TextArea } from "@components";
-import { STATUSES } from "@consts";
+import { STATUSES } from "@constants";
+import { statusControl, testErrorText, testInputPlaceholder } from "@mocks";
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, fn, userEvent, within } from "@storybook/test";
 import { type ChangeEvent, useState } from "react";
-import { statusControl, testErrorText, testInputPlaceholder } from "../mocks";
 
 const testValue = "Test Value";
 const testTextLimit = 100;
