+++ src/stories/Notification.stories.ts
@@ -1,8 +1,8 @@
 import { Notification } from "@components";
-import { SIMPLE_SIZES, STATUSES_NOTIFY } from "@consts";
+import { SIMPLE_SIZES, STATUSES_NOTIFY } from "@constants";
+import { statusControl, testErrorText, testSuccessText } from "@mocks";
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, within } from "@storybook/test";
-import { statusControl, testErrorText, testSuccessText } from "./mocks";
 
 const meta = {
 	title: "Base Components/Notification",
