+++ src/stories/StatusIcon.stories.tsx
@@ -1,8 +1,8 @@
 import { StatusIcon } from "@components";
-import { STATUSES } from "@consts";
+import { STATUSES } from "@constants";
+import { statusControl } from "@mocks";
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, within } from "@storybook/test";
-import { statusControl } from "./mocks";
 
 const meta = {
 	title: "Design resources/Icons/StatusIcon",
