+++ src/styles/components.css
@@ -70,3 +70,15 @@ input[type="search"]:focus::-webkit-search-cancel-button {
   opacity: 0.2;
   pointer-events: all;
 }
+
+input[type="range"]::-webkit-slider-thumb {
+  appearance: none;
+  width: 2px;
+  height: 10px;
+  border-radius: 2px;
+  @apply bg-white-100;
+}
+
+input[type="range"]:disabled::-webkit-slider-thumb {
+  @apply bg-transparent;
+}
