+++ src/styles/story.css
@@ -1,4 +1,4 @@
-@import "./export.css";
+@import "./index.css";
 @import "tailwindcss/base";
 @import "tailwindcss/components";
