+++ tsconfig.json
@@ -18,8 +18,10 @@
     "downlevelIteration": true,
     "paths": {
       "@components": ["./src/components"],
-      "@utils": ["./src/utils"],
-      "@consts": ["./src/consts"]
+      "@constants": ["./src/constants"],
+      "@mocks": ["./src/stories/mocks"],
+      "@styles": ["./src/styles"],
+      "@utils": ["./src/utils"]
     },
     "noUnusedLocals": true,
     "noUnusedParameters": true
